fix(Drive): compare with the target value threshold correctly
would be the same causing no process to occur which would then in turn
There was an issue where the AngularDrive treated the target value
threshold as actual degrees when compare with current pseudo rotation.
Instead, it should treat the target value threshold as normalized value.

This fix works out the equivalent threshold in degrees because this
usually results in a larger floating point value for stability.